Manually merged important changes in main and mike

+++ train_target_model.py
@@ -17,8 +17,8 @@ os.makedirs(Config.LOGS_PATH, exist_ok=True)
 os.makedirs(f'{Config.LOGS_PATH}/{Config.TARGET_MODEL_FOLDER}/', exist_ok=True)
 
 dm = MNISTDataModule(
-    f'{Config.LOGS_PATH}', 
-    batch_size=Config.TARGET_MODEL_BATCH_SIZE, 
+    f'{Config.LOGS_PATH}',
+    batch_size=Config.TARGET_MODEL_BATCH_SIZE,
     num_workers=Config.NUM_WORKERS,
     drop_last=True
 )
@@ -26,7 +26,7 @@ dm = MNISTDataModule(
 model = TargetModel()
 
 wandb_logger = pl_loggers.WandbLogger(
-    project='deep-learning', 
+    project='deep-learning',
     group='target_model',
     log_model=True,
     save_dir=Config.LOGS_PATH
@@ -34,23 +34,23 @@ wandb_logger = pl_loggers.WandbLogger(
 wandb_logger.watch(model)
 
 checkpoint_callback = ModelCheckpoint(
-    f'{Config.LOGS_PATH}/{Config.TARGET_MODEL_FOLDER}/', 
-    monitor = "val_loss", 
-    save_top_k = 1, 
-    save_last = True, 
+    f'{Config.LOGS_PATH}/{Config.TARGET_MODEL_FOLDER}/',
+    monitor="val_loss",
+    save_top_k=1,
+    save_last=True,
     mode='min'
 )
 
 callbacks = [checkpoint_callback]
 
 trainer = Trainer(
-    gpus=-1, 
-    max_epochs = 40, 
-    precision = 16, 
-    callbacks = callbacks, 
+    gpus=-1,
+    max_epochs=40,
+    precision=16,
+    callbacks=callbacks,
     benchmark=True,
-    num_sanity_val_steps = 2, 
-    logger=wandb_logger, 
+    num_sanity_val_steps=2,
+    logger=wandb_logger,
 )
 
 trainer.fit(model, dm)
